@@ -1,8 +1,8 @@
 package com.freenow.android_demo.activities;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.freenow.android_demo.App;
 import com.freenow.android_demo.dependencies.component.AppComponent;
