@@ -1,8 +1,11 @@
 package com.freenow.android_demo.adapters;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.VisibleForTesting;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.View.OnClickListener;
@@ -124,6 +127,11 @@ public class DriverAdapter extends ArrayAdapter<Driver> {
 
         void execute(Driver driver);
 
+    }
+
+    @VisibleForTesting
+    public List<Driver> getAllItems() {
+        return mDrivers;
     }
 
 }
